reorganize errros (#92)
